Extract fixture server host and port into constants
The constants are mirrored between the e2e logic (`e2e/`) and the app logic
in the ReadOnlyNetworkStore (`app/scripts/lib/network-store.js`) as they need
to be the same in both places. These could eventually be made dynamic but the
current build process isn't kind to that idea (in that the app is compiled
and run separately from the e2e tests, which aren't compiled at all).